Share the capture subscriber; pin the element-drop warn too
Extract the near-identical scoped-subscriber Recorders from wire.rs and
the new null-arm test into a #[cfg(test)] test_subscriber module with
capture_messages/capture_targets helpers, so the next warn-pinning test
is a one-liner instead of a third copy. The null-arm test folds into
lenient_vec_warns_are_not_silent, which now also pins the element-drop
warn — the other load-bearing signal in the helper, and the one that
fires on the two pre-existing envelopes whose Err-to-Ok
reclassification the CHANGELOG records. Also note on the sibling
helpers that their Some(Value::Null) pattern half is belt-and-braces
rather than reachable, so the next audit does not re-derive the 1c8b4a3
mistake.
