+++ src/lib.rs
@@ -57,6 +57,8 @@
 // =============================================================================
 pub(crate) mod http;
 pub(crate) mod serde_util;
+#[cfg(test)]
+pub(crate) mod test_subscriber;
 
 // =============================================================================
 // Core Type Modules
